fix: Modifying the lint logic to allow for similarly named projects
Right now, if a project has a dependency with the same starting root
name, it is picked as a dependency, but this may not true.

For instance, a project called weather, with a root folder of
'python/weather', has a dependency on a library in 'python/weather_lib'.
The startwith statement is going to call that a self-dependency, but it
is not the case.